Separating date validation to utils.
