+++ statsapiclient/schedule.py
@@ -1,11 +1,8 @@
 """This class allows retrieval of games from a given
 date or date range."""
 
-
-from datetime import datetime
-
-from statsapiclient.constants import SCHEDULE_PARAMS, SCHEDULE_DATE_FORMAT
-from statsapiclient.utils import fetch_json
+from statsapiclient.constants import SCHEDULE_PARAMS
+from statsapiclient.utils import fetch_json, validate_date
 
 
 class Schedule:
@@ -23,10 +20,10 @@ class Schedule:
         self.endpoint = "api/v1/schedule"
         self.params = {"expand": SCHEDULE_PARAMS}
 
-        self.validate_date(start_date)
+        validate_date(start_date)
 
         if end_date:
-            self.validate_date(end_date)
+            validate_date(end_date)
 
         self.params["startDate"] = start_date
         self.params["endDate"] = end_date if end_date else start_date
@@ -36,15 +33,3 @@ class Schedule:
 
         for date in json["dates"]:
             self.games += [game for game in date["games"]]
-
-    @staticmethod
-    def validate_date(date):
-        """Validates that a date meets the specified format."""
-        try:
-            datetime.strptime(date, SCHEDULE_DATE_FORMAT.get('format'))
-            return True
-        except ValueError:
-            format_display = SCHEDULE_DATE_FORMAT.get('display')
-
-            raise ValueError(
-                f"Incorrect date format, should be {format_display}")
