+++ statsapiclient/utils.py
@@ -1,8 +1,13 @@
 """Utility functions used across modules."""
+from datetime import datetime
 
 from requests import HTTPError, get
 
-from statsapiclient.constants import API_HOST, HEADERS
+from statsapiclient.constants import (
+    API_HOST,
+    HEADERS,
+    SCHEDULE_DATE_FORMAT,
+)
 
 
 def fetch_json(endpoint, params=None):
@@ -35,3 +40,15 @@ def fetch_json(endpoint, params=None):
         return response.json()
     except HTTPError:
         raise HTTPError
+
+
+def validate_date(date):
+    """Validates that a date meets the specified format."""
+    try:
+        datetime.strptime(date, SCHEDULE_DATE_FORMAT.get('format'))
+        return True
+    except ValueError:
+        format_display = SCHEDULE_DATE_FORMAT.get('display')
+
+        raise ValueError(
+            f"Incorrect date format, should be {format_display}")
